+++ gateway/tests/api/test_v1_program.py
@@ -13,7 +13,7 @@ from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
 from api.domain.authorization.function_access_result import FunctionAccessResult
-from tests.api.conftest import create_function_access_result
+from tests.utils import create_function_access_result
 from core.model_managers.job_events import JobEventContext, JobEventOrigin, JobEventType
 from core.models import (
     Job,
